test: update new api tests
